Added difficulties to TicTacToeAgent
Added the fllowing difficulties to TicTacToeAgent:

Easy: If there's a winning move, it goes for it.

Medium: Same as easy, but if the center tile is empty it goes for it 100% of the time.

Hard: Same as medium, but if it the opponent is winning the next turn, it tries to block it.